Update Controller check helpers' style
Helper functions for the check method: getRow, getColumn, and getSquare,
have been moved to SudokuBoard. The nature of these functions were more
appropriate in SudokuBoard rather than Controller.

Update the helper functions' style; such as replacing variables using
pot_hole style with the camelCase style, and removing unnecessary
whitelines that make it inconsistent with the rest of the code.

Fix a typo one of the helper function's docstring.