Add dialogue box for winning or losing the game
This commit adds 2 potential dialogue boxes to be displayed when the
check button is clicked: a winning message saying "Congratulations! You
have all the correct inputs!" will be displayed if the player has all
the correct inputs, and a losing message saying "Incorrect! Please
double check your inputs!" will be displayed if the player has the
incorrect inputs or if there are still emtpy cells to be filled.