Add style to Sudoku GridPane
Adds column and row margins to separate the inner 3x3 squares.

Increases the font size of the CellButtons to make it look more
eye-appealing.

Sets the font color of the CellButtons that cannot be modified by the
user to be black. This allow users to differentiate which cells are
modifiable by its font color.